+++ CHANGELOG.md
@@ -15,14 +15,14 @@ Emojis for the following are chosen based on [gitmoji](https://gitmoji.dev/).
 ### ✨ Features
 
 - Scribe-Data now has the ability to download Wikipedia dumps of any language.
-- Processes have been added to parse and clean the above dumps.
+- Functions have been added to parse and clean the above dumps.
 - Autosuggestions are generated from the cleaned texts by deriving most common words and those words that most commonly follow them.
-  <!-- - A query for profane words has been added and integrated into the autosuggest flow to make sure that inappropriate words aren't included. -->
+- A query for profane words has been added and integrated into the autosuggest flow to make sure that inappropriate words are not included.
   <!-- - The adjectives column has been removed from Scribe data tables until support is offered. -->
 
 ### ♻️ Code Refactoring
 
-- Update error messages for incorrect args in update_data.py
+- The error messages for incorrect args in update_data.py have been updated.
 
 # Scribe-Data 1.0.1